requirements.txt
flask
flask-cors
bcrypt
pyjwt
